user params object instead of passing separate params
